LayoutTests/imported/w3c/web-platform-tests/content-security-policy/securitypolicyviolation/blockeduri-inline-expected.txt


FAIL Inline violations have a blockedURI of 'inline' assert_equals: expected 9 but got 1
